refactored sidebar and navbars
