+++ client/src/components/NavBar/Logo.module.css
@@ -1,0 +1,5 @@
+/* logo */
+.logo {
+    height: 6rem;
+    margin-left: 2rem;
+  }
