+++ client/src/components/NavBar/NavBarItems.module.css
@@ -1,0 +1,47 @@
+/* user navigation */
+.user_nav {
+    align-self: stretch;
+    display: flex;
+    align-items: center;
+  }
+  .user_nav > * {
+    padding: 0 2rem;
+    cursor: pointer;
+    height: 100%;
+    display: flex;
+    align-items: center;
+  }
+  .user_nav > *:hover {
+    background-color: #f4f2f2;
+  }
+  .user_nav__icon_box {
+    position: relative;
+  }
+  
+  .nav__icon {
+    height: 2.25rem;
+    width: 2.25rem;
+  }
+  
+  .user_nav__notification {
+      font-size: .8rem;
+      height: 1.75rem;
+      width: 1.75rem;
+      border-radius: 100px;
+      color: #fff;
+      background-color: rgb(241, 70, 70);
+      position: absolute;
+      top: 1.5rem;
+      right: 1.1rem;
+  
+      display: flex;
+      justify-content: center;
+      align-items: center;
+  }
+  
+  
+  .nav__profile {
+    height: 3.75rem;
+    border-radius: 50%;
+    margin-right: 1rem;
+  }
