+++ client/src/components/UI/SideBar.js
@@ -1,43 +1,33 @@
 import React from "react";
 import classes from "./SideBar.module.css";
-import { FaHome, FaClipboardCheck, FaFolder, FaDonate } from "react-icons/fa";
+import {
+  FaHome,
+  FaClipboardCheck,
+  FaFolder,
+  FaDonate,
+  FaSignOutAlt,
+} from "react-icons/fa";
+import SideBarItem from "./SideBarItem";
 
 const SideBar = (props) => {
   return (
     <nav className={classes.sidebar}>
       <ul className={classes.side_nav}>
-        <li className={`${classes.side_nav_item}`} >
-          <a href="#" className={classes.side_nav_link}>
-            <svg className={classes.side_nav_icon}>
-              <FaHome size="2rem" color="#333" />
-            </svg>
-            <span>HOME</span>
-          </a>
-        </li>
-        <li className={classes.side_nav_item}>
-          <a href="#" className={classes.side_nav_link}>
-            <svg className={classes.side_nav_icon}>
-              <FaClipboardCheck size="2rem" color="#333" />
-            </svg>
-            <span>DASHBOARD</span>
-          </a>
-        </li>
-        <li className={classes.side_nav_item}>
-          <a href="#" className={classes.side_nav_link}>
-            <svg className={classes.side_nav_icon}>
-              <FaFolder size="2rem" color="#333" />
-            </svg>
-            <span>MY PROFILE</span>
-          </a>
-        </li>
-        <li className={classes.side_nav_item}>
-          <a href="#" className={classes.side_nav_link}>
-            <svg className={classes.side_nav_icon}>
-              <FaDonate size="2rem" color="#333" />
-            </svg>
-            <span>DONATION</span>
-          </a>
-        </li>
+        <SideBarItem name="HOME" href="#" >
+          <FaHome size="2rem" color="#333" />
+        </SideBarItem>
+        <SideBarItem name="DASHBOARD" href="#">
+          <FaClipboardCheck size="2rem" color="#333" />
+        </SideBarItem>
+        <SideBarItem name="PROFILE" href="#">
+          <FaFolder size="2rem" color="#333" />
+        </SideBarItem>
+        <SideBarItem name="DONATION" href="#">
+          <FaDonate size="2rem" color="#333" />
+        </SideBarItem>
+        <SideBarItem name="LOGOUT" href="#">
+          <FaSignOutAlt size="2rem" color="#333" />
+        </SideBarItem>
       </ul>
     </nav>
   );
